+++ sdk/storage/storage-blob/recordings/node/aborter/recording_should_set_value_and_get_value_successfully.js
@@ -1,3 +1,0 @@
-let nock = require('nock');
-
-module.exports.testInfo = {"container":"container156150780603601216"}