+++ sdk/storage/storage-queue/karma.conf.js
@@ -51,7 +51,7 @@ module.exports = function(config) {
     // inject following environment values into browser testing with window.__env__
     // environment values MUST be exported or set with same console running "karma start"
     // https://www.npmjs.com/package/karma-env-preprocessor
-    envPreprocessor: ["ACCOUNT_NAME", "ACCOUNT_SAS", "TEST_MODE"],
+    envPreprocessor: ["ACCOUNT_NAME", "ACCOUNT_SAS", "TEST_MODE", "STORAGE_SAS_CONNECTION_STRING"],
 
     // test results reporter to use
     // possible values: 'dots', 'progress'